--- PoExtractor.CS/Constants.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace PoExtractor.CS {
-    public static class Constants {
-        public static string[] StringLocalizerIdentifiers = new string[] { "T", "S", "TS", "H", "TH" };
-    }
-}
