+++ PoExtractor.Languages.CS/CSharpProjectProcessor.cs
@@ -1,0 +1,46 @@
+﻿using System.IO;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using PoExtractor.Core;
+using PoExtractor.Core.Contracts;
+using PoExtractor.Languages.CS.MetadataProviders;
+using PoExtractor.Languages.MetadataProviders;
+
+namespace PoExtractor.Languages.CS {
+    /// <summary>
+    /// Extracts localizable strings from all *.cs files in the project path
+    /// </summary>
+    public class CSharpProjectProcessor : RazorViewsProcessor {
+        public override void Process(string path, string basePath, LocalizableStringCollection strings) {
+            var codeMetadataProvider = new CodeMetadataProvider(basePath);
+            var csharpWalker = new ExtractingCodeWalker(
+                new IStringExtractor<SyntaxNode>[] {
+                        new SingularStringExtractor(codeMetadataProvider),
+                        new PluralStringExtractor(codeMetadataProvider)
+                }, strings);
+
+            foreach (var file in Directory.EnumerateFiles(path, "*.cs", SearchOption.AllDirectories)) {
+                if (Path.GetFileName(file).EndsWith(".cshtml.g.cs")) {
+                    continue;
+                }
+
+                using (var stream = File.OpenRead(file)) {
+                    using (var reader = new StreamReader(stream)) {
+                        var syntaxTree = CSharpSyntaxTree.ParseText(reader.ReadToEnd(), path: file);
+
+                        csharpWalker.Visit(syntaxTree.GetRoot());
+                    }
+                }
+            }
+
+            base.Process(path, basePath, strings);
+        }
+
+        protected override IStringExtractor<SyntaxNode>[] GetStringExtractors(RazorMetadataProvider razorMetadataProvider)
+            => new IStringExtractor<SyntaxNode>[]
+            {
+                new SingularStringExtractor(razorMetadataProvider),
+                new PluralStringExtractor(razorMetadataProvider)
+            };
+    }
+}
