+++ PoExtractor.Languages.CS/MetadataProviders/CodeMetadataProvider.cs
@@ -5,7 +5,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using PoExtractor.Core;
 using PoExtractor.Core.Contracts;
 
-namespace PoExtractor.CS.MetadataProviders {
+namespace PoExtractor.Languages.CS.MetadataProviders {
     /// <summary>
     /// Provides metadata for .cs code files
     /// </summary>
