+++ PoExtractor.Languages.VB/SingularStringExtractor.cs
@@ -3,12 +3,13 @@ using Microsoft.CodeAnalysis.VisualBasic;
 using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 using PoExtractor.Core;
 using PoExtractor.Core.Contracts;
+using PoExtractor.Languages;
 using System.Linq;
 
-namespace PoExtractor.VB
+namespace PoExtractor.Languages.VB
 {
     /// <summary>
-    /// Extracts <see cref="LocalizableStringOccurence"/> with the singual text from the VB AST node
+    /// Extracts <see cref="LocalizableStringOccurence"/> with the singual text from the C# & VB AST node
     /// </summary>
     /// <remarks>
     /// The localizable string is identified by the name convention - T["TEXT TO TRANSLATE"]
