+++ PoExtractor.Languages/RazorPageGeneratorResult.cs
@@ -1,0 +1,7 @@
+﻿namespace PoExtractor.Languages
+{
+    public class RazorPageGeneratorResult {
+        public string FilePath { get; set; }
+        public string GeneratedCode { get; set; }
+    }
+}
