+++ PoExtractor.Languages/ViewCompiler.cs
@@ -1,24 +1,27 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Extensions;
 
-namespace PoExtractor.CS {
-    static class ViewCompiler {
-        public static IList<RazorPageGeneratorResult> CompileViews(string projectDirectory) {
+namespace PoExtractor.Languages
+{
+    public static class ViewCompiler
+    {
+        public static IList<RazorPageGeneratorResult> CompileViews(string projectDirectory)
+        {
             var projectEngine = CreateProjectEngine("PoExtraxtor.GeneratedCode", projectDirectory);
 
             var results = new List<RazorPageGeneratorResult>();
 
             var viewDirectories = Directory.EnumerateDirectories(projectDirectory, "Views", SearchOption.AllDirectories);
-            foreach (var viewDir in viewDirectories) {
+            foreach (var viewDir in viewDirectories)
+            {
                 var viewDirPath = viewDir.Substring(projectDirectory.Length).Replace('\\', '/');
                 var viewFiles = projectEngine.FileSystem.EnumerateItems(viewDirPath);
 
-                foreach (var item in viewFiles.Where(o => o.Extension == ".cshtml")) {
+                foreach (var item in viewFiles.Where(o => o.Extension == ".cshtml"))
+                {
                     results.Add(GenerateCodeFile(projectEngine, item));
                 }
             }
@@ -26,9 +29,11 @@ namespace PoExtractor.CS {
             return results;
         }
 
-        static RazorProjectEngine CreateProjectEngine(string rootNamespace, string projectDirectory) {
+        public static RazorProjectEngine CreateProjectEngine(string rootNamespace, string projectDirectory)
+        {
             var fileSystem = RazorProjectFileSystem.Create(projectDirectory);
-            var projectEngine = RazorProjectEngine.Create(RazorConfiguration.Default, fileSystem, builder => {
+            var projectEngine = RazorProjectEngine.Create(RazorConfiguration.Default, fileSystem, builder =>
+            {
 
                 builder
                     .SetNamespace(rootNamespace)
@@ -44,19 +49,16 @@ namespace PoExtractor.CS {
             return projectEngine;
         }
 
-        static RazorPageGeneratorResult GenerateCodeFile(RazorProjectEngine projectEngine, RazorProjectItem projectItem) {
+        public static RazorPageGeneratorResult GenerateCodeFile(RazorProjectEngine projectEngine, RazorProjectItem projectItem)
+        {
             var codeDocument = projectEngine.Process(projectItem);
             var cSharpDocument = codeDocument.GetCSharpDocument();
 
-            return new RazorPageGeneratorResult {
+            return new RazorPageGeneratorResult
+            {
                 FilePath = projectItem.PhysicalPath,
                 GeneratedCode = cSharpDocument.GeneratedCode,
             };
         }
-    }
-
-    class RazorPageGeneratorResult {
-        public string FilePath { get; set; }
-        public string GeneratedCode { get; set; }
     }
 }
