+++ PoExtractor.OrchardCore/PoExtractor.OrchardCore.csproj
@@ -21,9 +21,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PoExtractor.Core\PoExtractor.Core.csproj" />
-    <ProjectReference Include="..\PoExtractor.CS\PoExtractor.CS.csproj" />
+    <ProjectReference Include="..\PoExtractor.Languages.CS\PoExtractor.Languages.CS.csproj" />
     <ProjectReference Include="..\PoExtractor.Liquid\PoExtractor.Liquid.csproj" />
-    <ProjectReference Include="..\PoExtractor.VB\PoExtractor.VB.csproj" />
+    <ProjectReference Include="..\PoExtractor.Languages.VB\PoExtractor..Languages.VB.csproj" />
   </ItemGroup>
 
 </Project>
