+++ PoExtractor.OrchardCore/Program.cs
@@ -3,9 +3,9 @@ using OrchardCore.DisplayManagement.Liquid.Tags;
 using OrchardCore.DynamicCache.Liquid;
 using PoExtractor.Core;
 using PoExtractor.Core.Contracts;
-using PoExtractor.CS;
+using PoExtractor.Languages.CS;
 using PoExtractor.Liquid;
-using PoExtractor.VB;
+using PoExtractor.Languages.VB;
 using System;
 using System.Collections.Generic;
 using System.IO;
